pyproject.toml
[tool.poetry]
name = "pgid"
version = "0.0.0"
description = ""
authors = ["Orlando Hohmeier"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-benchmark = {extras = ["histogram"], version = "^4.0.0"}
pytest-docker-tools = "^3.1.3"
ruff = "^0.4.9"
asyncpg = "^0.29.0"
pytest-asyncio = "^0.23.7"
psycopg2-binary = "^2.9.9"
ulid = "^1.1"
pip = "^24.0"
install = "^1.3.5"
uuid7 = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
